[File Mgmt] Drag/drop NIF files to window to open
Allows one to open another NIF (or BTO, BTR, etc.) file in the current
window.  Dragging multiple files will open a new window for each file,
unless the window you are dragging them to a blank window.  In that case
the first file in the list will be opened in the blank window and a new
window will be created for each other file.