[Build] Multiplatform build improvements
If a CONFIG was not using "debug_and_release", large swaths of the
NifSkope.pro file were being ignored because `build_pass` relies on
"debug_and_release" mode being set.

It seems on Linux that "debug_and_release" is not the default as on
Windows.  So, "DESTDIR" wasn't being set at all on Linux (or for anybody
who turns off "debug_and_release") and was causing lots of issues.

Also improved the mkdir behavior across platforms for the post-build
file copies.