[GL] Shader: Environment Mapping fixes
When an environment mask was not present, the normal map was not being
used instead.

Expanded to apply to eye environment mapping flags.

Also, environment mapping would not turn off if
SLSF1_Environment_Mapping was not checked, which could lead to confused
users.  Both the shader type and the flag need to be set in order to see
reflections.