[GL] Renderer: Support ZBuffer flags
Support for SLSF1_ZBuffer_Test and SLSF2_ZBuffer_Write which is a huge
improvement for rendering quality as most alpha effects before were
being written to the depth buffer still, and would cause weird cut-out
effects with other geometry.

Also, generally supporting these flags means it is obvious when your
solid mesh does not have the flags on.  It is possible NiAlphaProperty
with blending overrides this, but will have to test.