[Settings] New Settings classes
SettingsDialog - The actual dialog window, includes a list widget to
list the pages and a stacked widget to house the pages.
SettingsPane - Abstract class for each "page" of the settings dialog,
other Settings* classes inherit this and refer to a specific .ui file.

When values in a SettingsPane are modified, they change the status of
the pane to "modified" so that things like Save/Apply can be
greyed/ungreyed, and also when writing the settings only the panes that
are modified need to write their settings.

The SettingsDialog emits several signals during Save/Apply, which other
classes can listen for to update their local copy of the settings.