[Fix] Several fixes to migration
There were several issues with migrating from dev1 to dev2.  Namely a
crash was guaranteed to occur unless a certain value was set by hand in
the registry first.

Also, even after doing so, default settings were not getting written the
first time due to changes in the way write() relies on isModified().

Furthermore the "AUTO" value from the previous FSEngine storage needed
to be expanded in place to the actual list of files.