[Build] Necessary fixes for compiling with Qt 5.5
A number of things changed in Qt 5.5, especially centered around OpenGL.

Qt now does dynamic switching between OpenGL or ANGLE instead of it
being an option when building Qt.

Build step was modified to avoid some unnecessary DLLs from being copied
(which are new as of Qt 5.5).

An explicit include for QDataStream became necessary for whatever
reason.

The additional fix had to do with the new GCC version not liking my C++
usage.