[Select] Initial Selection Modes commit
Can enter Select Vertex mode and select vertices from the viewport,
works for both FO4 and pre-FO4 meshes.  There is an internal limitation
of 256 TriShapes due to the way the color encoding in the color buffer
used for vertex selection works. This is also taking into account the
limit on vertices is 65535 for any one trishape.   This makes for a
total limit of ~16M vertices (256^3).