[XML] Change nifskopetype="blob" to binary="1" for rows
The "nifskopetype" attribute could also be used on rows to create
aliases, when wanting to handle the data differently than the type=""
attribute described.

This was only used to treat some arrays of bytes like blobs though, so
instead define these rows as binary which is project agnostic and
further helps to clarify the data definitions for those rows.