[XML] Initial Support for "externalcond" attr
The "externalcond" attribute is a contract for compounds which says that
cond="" attributes on its rows are resolved externally to the compound.
This means that its conditions cannot reference other rows, and can only
come from ARG passing.

This is useful for optimizing arrays of compounds where the conditions
for each row of the compound will be the same between compounds,  and as
a contract is telling the XML reader that the rows in this compound
cannot rely on other rows for their conditions.

This commit is just the basis, the usage will come in later commits.