[NifModel] Refactor
In an effort to speed up and also simplify things several changes were
made.  First "fast" params/checks were removed from all the loading
code.  This was done to prevent the view from updating as the file was
loaded which slowed things down.  See previous commit for why this was
unnecessary.

Second introduced a state stack to the model to track what the model is
currently doing.  The state is used in various places to prevent the
emission of signals.

A few method names were also changed for legibility. (Too many "load"
and "save" methods)

Lastly all spells cast put the NifModel into a specific state which
prevents all dataChanged signals and instead stores whether or not there
was any data changed when the spell was cast.