[NifTreeView] Row Hiding refactor
Redoes the row hiding based on conditions.  The previous only hid
version mismatched rows, this hides everything non-applicable to the
block and heavily reduces visual clutter.

It also removes the "realtime" option as it is always "realtime" by
default given the NifModel refactor and the ability to prevent row
condition updating when necessary.