[GL] TexCache file watcher fixes
If the `tx->reload() = true; emit sigRefresh();` branch was triggered
the data stored for the texture would not be updated.  Depending on how
fast the file watcher emits the signal the texture may or may not
refresh because the results of QFile::exists() would vary.

Also prevented this code from running more than once for the same
texture which it was doing.