[GL] Prevent multiple successive mesh recompiles
Sometimes NifModel::linksChanged can be emitted dozens or hundreds of
times right after the other.  This caused GLView to attempt an update
for each call.  This timer limits the recompile to once every 16ms at
most, which is just under the frametime at 60FPS.  However if there are
100 calls each 10ms apart the timer will only timeout and call
modelLinked() once,  16ms after the 100th call.