[UI] Copy Transform button in Inspect
Transform > Copy in the context menus only allows for copying of the
local transform of the block.  However the world transform is sometimes
useful for copying, like the world transform of bones in a skeleton NIF.
This is particularly necessary when creating Art Object NIFs which
attach to bones.

The world transform is inaccessible via Spells, so I added a button to
Inspect.