[UI] Remove Name editing from NifBlockEditor
The standard base editor adds the editing of Name which means it shows
up in things like Edit > Transform.  This isn't thematically appropriate
considering the window should just be for editing the Transform.

It also had no time delay in editing the header with the new string, so
a string was added for each subsequent character you type.