[NifExpr] Division and multiplication for expressions
It became necessary to divide two numbers to get an array size in recent
XML changes.  This requires a symbol for division to be parsed by the
Expression class.  However just `/` is not good enough because it can
occur too often in normal strings, and it's also inserted as part of
`../` for child rows' arr1, etc.