[NifModel] Introduce mixins (specialized compounds)
A series of XML updates from ttl269 broke many features of the program
because they introduced unneeded nesting.  Their use was solely to
reduce repetition and not to show hierarchy so I am now treating them as
if they are includes.

This works almost identically to the way abstract niobject work except
the mixin can be anywhere in the block and not only the front.  I will
likely propose this be added to the XML itself as it denotes usage of
the compound.