[UI] Fix checkbox list crash
Finally determined after many years the reasoning behind the random and non-reproducible crash when editing bitflags.   `editor->setGeometry`  in `CheckBoxListDelegate::updateEditorGeometry` would run on each item hovered on but only sometimes did this crash.  It turns out it is because of the item view having auto-scroll on by default.  Trying to call setGeometry with auto-scroll on could crash randomly.

Threw in some null checking for good measure though it was unnecessary.