[Spells] Transform > Apply fixes, support for BSTriShape
Transform > Apply never worked correctly when a mesh had Tangents and Bitangents and the Transform had a non-identity rotation.  Tangents/Bitangents were not transformed so you would need to run Update Tangent Space to correct this.

Also added support for BSTriShape.  The location of skinned Vertex Data being in the partition for SSE does not factor in since Apply should not be run on skinned geometry anyway.