[GL] Fix rendering issues
glLightModeli appeared to be causing GPU panic in NiMesh NIFs when zooming in.  FPS would plummet and GPU usage would be at 99.9%.

get_max_anistropy() was inadvertently being called every frame for every mesh for every texture.  This was causing massive slowdowns in large NIFs with thousands of blocks.

pow() in setSamples() was reversed so samples was always artificially high past 0 or 1.

After changing to Qt5 OpenGL, DrawTriangleIndex no longer works but the call causes major slowdowns when used.

Removed index lookups for names that no longer exist and short circuit the draw functions when the rest of the block is Bethesda-only for non-Bethesda NIFs.