[NIF] 20.3.1.2 custom RTTI hash support
There are some games with a custom NIF version that has no RTTI strings in the header (Block Types) but instead a list of hashes.  It was fairly nonintrusive to implement support for this format as you simply look the block up by hash instead of by name.  The hashes are added when the XML is read.  The hash used is a version of djbhash.