[GL] Skinning Improvements
The geometry displayed for skinned NiTri-based objects always came from the triangles on the shape.  However for shapes with a partition, the Triangles array is completely optional on the shape.  This is present in some Skyrim NIFs, so these NIFs do not render.

These changes make it so that the geometry will switch between the shape geometry and the partition geometry with the Skinning toggle.  Now any shapes with only partition geometry will show up, and correctly disappear when turning off Skinning.

This will also show any problems with partitions that might go uncaught because it was showing the wrong geometry.

Also adds a spell to update the shape triangles from the partition triangles.