[GL] Correct 20.1+ UV Set selection, 20.5+ Anisotropy for NiTexturingProperty
The UV index for maps in NiTexturingProperty was moved to the flags value's lower byte.  The code was only dealing with the clamp and filter modes in the upper byte.  This broke light maps and other secondary maps that used multiple texcoords.

Moved anistropy management to glproperty and added support for Max Anisotropy value in TexDesc.  It will now override the Anisotropic Filtering user setting for that texture.

Also modified the Flags UI for TexDesc to more easily edit the UV Index.