nifxml 0.9 sync, manual externalcond for BSVertexData
Since BSVertexDesc is an internal type, need to special-case the arg attribute for the Vertex Desc.

Also forced `externalcond` on BSVertexData as it's still not part of the XML spec and that way I don't have to worry about including it in the packaged XML for FO4/SSE to not be brutally slow.