[GL] NiMesh rewrite No. 3
There were a multitude of issues with my previous implementation, though not as severe as the initial code.  It was adding triangles for each datastream among other things.  Submeshes were also not supported.

Currently an index can still not exceed USHRT_MAX so if the index in a submesh + the offset is allowed to exceed this, then additional changes will need to be made, such as actually creating Mesh child classes for each submesh.

Skinning is also still not supported.