[UI] Add New Ref feature for "None" Ref links
Adds a button to all None Refs that allows you to insert a new block of the proper type.

Has filtering by type, version, and for controllers and interpolators limits it to the valid blocks for the parent or target type.

Replaced the menu flyout creation for the other insert operations and improved the categorization and filtering on it using the new filtering methods.

Also improved Bethesda NiExtraData auto-naming for new block insertions.