[GL] Renderer refactor, stage 1
Changed nearly all of setupProgram for the shaders.  Store the uniform locations, use enums and maps instead of strings to refer to the uniforms.  This helps with speed, legibility, and autocompletion.

Removed all the `ob_` shaders for the time being as there is a shader complexity issue (more shaders == worse).  Will reinstate fo3 and ob shading over time and with far fewer shaders.