[UI] Value string changes
Split the enum bitflag options after a user-defined number of options.  Currently WIP.  The combo box's line edit disables the newlines when the editor is created and reverts back to single line while the row stays the same height.

Also show "undefined" when a bool value is 2, which is the undefined value for NiBool.