[GL] Fix shader issues 2
Continuation of df48fc7

Do not return false for any missing textures, instead always assure binding and proper defaults.

Created option to show error color for missing diffuse.  This does not apply to BSESP as diffuse is not necessary for those.

Correctly blacken BSTriShape mesh with mismatched vertex attributes and shader flags for vertex colors.

Correctly show vertex colors for FO4 BSLSP if they exist in the vertex data.  The vertex color flags are optional since they do not exist in the BGSM.