Initial Game Manager commit
Resource collisions between Fallout 4 and Fallout 76 were severe due to a vastly different asset pipeline while maintaining the same filenames.  Technically resource collisions happened with Skyrim and Skyrim SE too, but the vanilla files were largely identical.

The game manager associates texture and material lookups to a particular game, based on the NIF version.

This does mean that Fallout 3 and Fallout NV are considered the same meta-game and lookups for either game's files will be done in both games' paths.

A "game mode override" is planned to address this.