[GL] Updated CRC to flag maps
Re-scanning the latest files showed a missed `NO_EXPOSURE` usage. Unsure if there since release or if added.  `LOD_OBJECTS` was also missing in NifSkope but was already present in nifxml.  Outside of BTOs, one single NIF uses `NO_EXPOSURE`.

As of August these are the only used shader flag CRCs.

Since editing of this version will not be supported, unused flags will continue to be excluded.