[UI] Message improvements
Return pointer from Message helpers when needing to customize the behavior.

Added a log message function directly to the model which allows any function with a nif pointer to support both UI messages and the file checker output.  For example, functions that are sanitizers (run on save) and used to output information in the file checker.