Game Manager update
Switched back to absolute paths for game folders, fixed auto detection in the UI and at initializing of said paths.

The idea was to show only relative paths for both folders and archives if they were under the install path as this would tidy up the UI.  In practice, supporting both relative and absolute (outside of game install) paths in the same list in the UI and registry is not worth the effort.