Fix: cubic key interpolation uses the derivatives
The interpolation request declares a type "T", which will be used in the
getter methods. When "value" is a Vector3, the getters will get
Vector3s; when "value" is a float, the getters will get floats. This
procedure therefore works with 3D interpolation (e.g. translocation)
as well as 1D interpolation (e.g. scale).